Organize os_updater files
