+++ internal/inbd/server.go
@@ -9,7 +9,7 @@ import (
 	"net"
 	"os"
 
-	osupdater "github.com/intel/intel-inb-manageability/internal/os_updater"
+	emt "github.com/intel/intel-inb-manageability/internal/os_updater/emt"
 	"google.golang.org/grpc"
 )
 
@@ -52,11 +52,13 @@ func RunServer(deps ServerDeps) error {
 	grpcServer := deps.NewGRPCServer()
 	deps.RegisterService(grpcServer)
 
-	// VerifyUpdateAfterReboot verufies the update after reboot.
+	// VerifyUpdateAfterReboot verifies the update after reboot.
 	// It compares the version in dispatcher_state file with the current system version.
 	// If the versions are different, the system successfully boots into the new image.
-	// TODO: Remove the hardcoded "EMT", and use DetectOS to get the OS type.
-	err = osupdater.VerifyUpdateAfterReboot("EMT")
+	
+	// TODO: This should call the correct function using the factory pattern.
+	// Currently hardcoded to use EMT.  It should also be able to call Ubuntu.
+	err = emt.VerifyUpdateAfterReboot("EMT")
 	if err != nil {
 		return fmt.Errorf("[Post verification failed] error verifying update after reboot: %w", err)
 	}
